--- src/main/java/duke/dukeException/DukeException.java
@@ -1,4 +1,0 @@
-package duke.dukeException;
-
-public class DukeException extends Exception{
-}
