--- src/main/java/duke/dukeTask/Deadline.java
@@ -1,16 +1,0 @@
-package duke.dukeTask;
-
-public class Deadline extends Task {
-
-    protected String by;
-
-    public Deadline(String description, String by) {
-        super(description);
-        this.by = by;
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + by + ")";
-    }
-}
