--- src/main/java/duke/dukeTask/Event.java
@@ -1,16 +1,0 @@
-package duke.dukeTask;
-
-public class Event extends Task {
-
-    protected String at;
-
-    public Event(String description, String at) {
-        super(description);
-        this.at = at;
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (at: " + at + ")";
-    }
-}
