--- src/main/java/duke/dukeTask/Task.java
@@ -1,23 +1,0 @@
-package duke.dukeTask;
-
-public class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public String getStatusIcon() {
-        return (isDone ? "[\u2718]" : "[ ]"); //return tick or X symbols
-    }
-
-    public String toString(){
-        return getStatusIcon() + description;
-    }
-
-    public void markAsDone() {
-        isDone = true;
-    }
-}
