--- src/main/java/duke/dukeTask/Todo.java
@@ -1,13 +1,0 @@
-package duke.dukeTask;
-
-public class Todo extends Task {
-
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
